Added ${ARFLAGS} to provide flags for ${AR}.
configure now sets a default ABI to use rather than THE ABI that will be
used when compiling.  Makefile.base allows the use of this default, or it
can be overridden on the command line.  Any compiler flags needed to
compile with a given ABI are set by Makefile.base.  Possible ABIs are:

	O32 => -o32 -mips2
	N32_M3 => -n32 -mips3
	N32_M4 => -n32 -mips4
	64_M3 => -64 -mips3
	64_M4 => -64 -mips4

${OGLROOT} and ${PFROOT} are once again substituted by configure though
no Makefiles are currently using those variables.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@169 08b38cba-cd3b-11de-854e-f91c5b6e4272