Separate the ABI and the instruction set with a new variable $INST_SET.
Prevent the expansion of some variables that are substituted into the
Makefiles until they are actually used in the Makefiles.  This is
especially useful when overriding the default ABI that results in a
change to the library directory suffix.

Added $TEST_VJ_LIBS so that the test code can get its library from the
appropriate lib directory including the ABI-derived suffix.  Related to
this and the above change, $TEST_VJ_LIBS_SUBST and $TEST_EXTRA_LIBS_SUBST
were added.  These are used when generating the Makefiles at the time of
installation.  Parts of these variables are prevented from expanding until
that time.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@173 08b38cba-cd3b-11de-854e-f91c5b6e4272