Added 'all-abi' and 'install-all-abi' targets for building every supported ABI on the target platform and installing them respectively.
With this comes better support for building multiple ABI/instruction set
combinations.  The object files are now built in an 'obj' directory tree
created at compile time that separates the objects by target platform,
ABI, instruction set and debugging/optimization flags used.  Each
category gets a separate directory in the tree, and the objects are built
at the leaves.  The libraries are now built in lib* subdirectories of the
top-level build directory rather than in the same directory as the object
files.  Thus, the only pollution in the source directory is the link to
config.h currently needed for the $VJ_BASE_DIR and instlinks stuff.

The order of the dependencies in the 'all' and 'install-all' targets has
been changed.  Now the debugging versions are built last thus making them
them the default versions to which the symbolic links made in the lib*
directories point.  These links are made when the library binaries are
created rather than by the 'links' target.

The 'info' target has had the default instruction set and supported ABI
list added to its output.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@174 08b38cba-cd3b-11de-854e-f91c5b6e4272