Build the Performer and OpenGL API code in separate libraries.  They are named libJuggler_pf and libJuggler_ogl respectively.
The recursive build process has changed slightly to accomodate this.
${OBJDIR} is now overridden by a ${BASE_OBJDIR} variable that gives the
top-level directory where the object files are built.  This allows a
subdirectory to be appended to provide for object files to be built in
separate directories.  This is how the GL and Performer APIs are
separated from the rest of the VR Juggler library.

For the test code, the libraries linked are now separated to include only
those necessary for a given application.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@249 08b38cba-cd3b-11de-854e-f91c5b6e4272