Added the capability to test a class to see if it has any member variables. The format is:
  {if membervar [!][type]}
  {endif}

The "type" is an optional argument that can be one of public, private,
protected or member.  If this argument is given, the test will only return
true if the type of member variable requested exists.  It may also be
negated in which case it will return true if member variables exist that
are not of the type named.  For example, if type is 'private' and it is
negated, then the test will return true if the class contains public or
protected member variables.  If the type is 'member', then it will return
true if the class has any member variables.

This change added the subroutines isvar() to test if a member is a
variable and getMemberVars() to return the list of a class' member
variables of a certain type.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@540 08b38cba-cd3b-11de-854e-f91c5b6e4272