Added the capability to test a class to see if it has any methods.  The format is:
  {if method [!][visibility]}
  {endif}

The "visibility" is an optional argument that can be one of public,
private, protected or member.  The behavior of this second argument is the
same as for the "membervar" test.

This change added the subroutine getMethods() to return the list of a
class' methods of a certain visibility.

It also backs out code that was accidentally committed in Rev. 1.12.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@542 08b38cba-cd3b-11de-854e-f91c5b6e4272