More event-system overhaul - adds chunkDBListener stuff, for listening to changes to a specific chunkdb.
moves chunkdbtreemodels into a cache in the ChunkDBPanel, where
it belongs, removing a lot of Core's knowledge of the GUI.
Introduces the db treemodel factory, which creates and caches the
treemodels and jtrees that are displayed in the ChunkDBPanels.

BUG: current code will let you view the same db in both sides of
the configure pane.  DO NOT MODIFY any chunks while you're doing
this... you'll head into an infinite recursive event-notification
loop.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@1214 08b38cba-cd3b-11de-854e-f91c5b6e4272