.vjconfig directory was really stressed before.  Now it is only recommended for advanced users.
This will help out sysadmins who want everyone to use the
"system" configuration (i.e. juggler/data/configfiles)

It will also help reduce redundent copies of the config files.

On the page, users are encouraged to find out their local site policy on
these files - and copy them to their own directory if there is none, or if
they are "advanced" and want to be able to change their simulator setup.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@1860 08b38cba-cd3b-11de-854e-f91c5b6e4272