BasicConfig application works.    see the terrain.config and the .dsc file for details
It's cool, I was able to runtimereconfigure the models and sounds.
models can be repositioned, scaled, changed from collidable to not, etc...
sound can be positioned, or set to ambient (non-positional) sound.

The haunted mansion app was up and running with new models and sounds in 3
minutes. :) yay.

I added a fix for ambient sounds not starting.

I added subgraph for sound, and one for noncollidable nodes.
when reconfiguring, the app nukes it's old subgraph (for sound or model
depending), then recreates it from the new description.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@2191 08b38cba-cd3b-11de-854e-f91c5b6e4272