Doing
    $gid = (getgrnam("$gname"))[2] or die ...;

is a bad situation when the returned gid is 0.  Perl sees that 0 and
thinks that getgrnam() failed when it actually succeeded.  To fix this
case, the array returned by getgrnam() is now stored in a temporary array,
and then the gid is extracted from that array.  The same has been done for
getpwnam().


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@2455 08b38cba-cd3b-11de-854e-f91c5b6e4272