The InstallWeb module has been broken up into the new object-oriented modules WebFile, HtmlFile and XmlFile.  WebFile is a base class that contains all the operations that can be performed on files using.  WebFile is also where all the commands are defined.  HtmlFile and XmlFile inherit from WebFile and define their own filter() methods (formerly InstallWeb::htmlFile() and InstallWeb::xmlFilter() respectively).
WebFile is a resuable module that does not necessarily have to be tied to
install-web since part of its initialization process sets the command
string using a user-defined value (which happens to default to
'install-web').  More subclasses similar to HtmlFile and XmlFile can be
added as necessary thus allowing us to extend the functionality of
install-web or other utilities that might benefit from the work already
done here.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@3195 08b38cba-cd3b-11de-854e-f91c5b6e4272