Fixed up the VR Juggler installation so that we can install in a directory such as /usr/local and get along with the other packages already installed. More specifically, all the extra Juggler stuff such as the sample applications and the Data directory tree go into the general data directory (usually ${prefix}/share).
This is not a complete solution because what we really want is to install
these extra files in ${prefix}/share by default but to allow that path to
be changed when the build is configured.  For example, if we install in
/usr/local, we want the extra files to go into /usr/local/share/vrjuggler
with $VJ_BASE_DIR set to /usr/local.  Unfortunately, there are many places
where ${VJ_BASE_DIR}/share... is hard-coded, and I haven't come up with
a good way yet for the configure script to alter that path.


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@3709 08b38cba-cd3b-11de-854e-f91c5b6e4272