COPYRIGHT COMMENTS: the copyright comments were buried in the middle of this file
OUPUT:
added output for number of tests that fail.


FAILED TEST (did not fix):
there is one test that fails, but is it important?? i mean we've been
using quat for a long time now, maybe the gl_quat is wrong?

the failed test is testing the setting of a mat from a gl_quat and from a
vj_quat.

what is wrong is that the resulting of gl_mat is transposed from
the resulting vj_mat (rotation component)

so maybe it is a formatting issue where checking the data is being done
columnwise instead of row wise, or vice versa...


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@4095 08b38cba-cd3b-11de-854e-f91c5b6e4272