Added a driver for the Polhemus Fastrack 3D system.  At this time, the driver only works on IRIX with SPROC threading, but that will change at some point in the future.
Note: Under the release model that I have been advocating for VR Juggler,
this is not how a driver, or any new feature, should be added (i.e., to the
release branch first).  In this case, however, we are strapped for time;
the Input Manager in the HEAD branch has diverged significantly; and the
author has been waiting for a long time for this to be added.  Thus, I'm
willing to make an exception.  :)

    Submitted by: Nicolas Tarrin <nicolas.tarrin@inria.fr>


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@4257 08b38cba-cd3b-11de-854e-f91c5b6e4272