Okay,
All of these files have been updated for Mac OS X support. Hopefully, if
all goes well, there should be no problems in compiling on currently
supported OS's.  There are a couple support Libraries and frameworks that
are needed for this to compile and operate.  Hopefully I can get around to
making a script to do this for the user.

One note, there was an incompatibility with the OS X Libraries and the
T_INVALID enum declared in Config/vjVarValue.h... This has been changed to
VJ_T_INVALID.

BuildSystem/config/sys.m4 has been updated to include darwin1.3 as a viable
OS type, as with configure.in.  The rest is pretty much straight-forward
cross-platform compatibility modifications.

6 source files have been added to manage sim input windows and GL surface
displays.

The libraries which need to be included for this to compile are:
NSPR 4.0
libftime.a <-- This is a library which I threw together to add compat.
SetupGL.framework <-- some functions stolen from Apple example code for
                      setting up GL contexts.  This can eventually be
                      merged into Juggler so that this framework is not
                      needed.

Some symbolic links must be made as well.
/usr/includes/GL needs to point to /System/Library/Frameworks/OpenGL.
framework/Headers
All of the GL libraries need to have symbolic links located in /usr/lib

Ben Thompson


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@4262 08b38cba-cd3b-11de-854e-f91c5b6e4272