Fixed up a couple of crashing threads. Increased speed by removing some iterations, reducing sleeps, and re-introducing .join() funcs.
Note: This means that it will not compile and run on sproc now.  This is good (the test is telling us that sproc is broken).  We should fix it in the future.
Note2: Because of all the sleep's in the tests, they are definately unstable.  When I write the acceptor connector tests I will try to write an example test that shows how to use mutexes,
semaphores, and conditions variables so that sleep-fudging is not necessary.  This should make the tests more stable and much faster to execute.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@4516 08b38cba-cd3b-11de-854e-f91c5b6e4272