Added support for creating and installing a vpr-config script.  This script can be used by other software to get compiler and linker flags needed for using VPR.
This version of the script is based on gtk-config.  Doing it this way
turned out to be very unpleasant since GTK+ is not written to be used on
both UNIX and Win32 systems.  Because of that, this iteration of support
is extremely sloppy and *should not* be used as a basis for similar efforts
in the future.  Once all everything calms down, I will come back to this
and improve it.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@4524 08b38cba-cd3b-11de-854e-f91c5b6e4272