Enable the button data for each bird by default.  I have tested this, and everything appears to be working fine.
For us to take advantage of the button data, we need to configure it for
and read it from each individual bird.  The driver (technically) allows
this, but the configuration does not at this time.  There are many settings
that can be configured for each bird, and we need to come up with a way of
doing that.  It will not happen, unfortunately, without breaking the chunk
format for the MotionStar.


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@4674 08b38cba-cd3b-11de-854e-f91c5b6e4272