Fixes run-time crash in the initialization of the vjKeyboardWin32.  Crashes in samplem_keys() seemed to be due to an implicit static cast from vjKeyboardWin32* to vjKeyboard* in startSampling().
The change that resulted in this problem is likely the modification of
vjKeyboardWin32's declaration from:
     class vjKeyboardWin32 : public vjKeyboard
to:  class vjKeyboardWin32 : public vjInput, public vjKeyboard
which changed its layout in memory.

Perhaps this should be looked at more closely since changing the
order of inheritance to:
   class vjKeyboardWin32: public vjKeyboard, public vjInput
also seems to cause incorrect copying/assignment of pointers.
If we could avoid the need to use a C-style cast, that would probably fix
this problem.  Unfortunately, that doesn't seem very reasonable the
way the object needs to be passed into a thread.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@4678 08b38cba-cd3b-11de-854e-f91c5b6e4272