Enhanced support is now provided for directly referring to embedded ConfigChunks.
For example, the Display Window chunk in Juggler 1.1 can embed a
Keyboard chunk; keyboard proxies can now refer directly to this
embedded chunk.  The syntax looks like
"chunkname->property_token->embeddedchunkname".  The name of the
embedded chunk itself contains _the entire path_ as shown above.
(the separator symbol "->" was selected because i was afraid the
other candidates - '/', '.', ':', etc., might be useful inside
chunk names, but i'm open to suggestions/arguments. this is easily
changed, but will munge config files that use the feature).

VjControl was enhanced in the following ways:
- embedded chunks are renamed to include the full path when a config
file is loaded.  the chunk editor panel only displays the trailing
element of the name, since that's the only part that can reasonably
be modified.
- the file verification feature can now check for dependencies on
embedded chunks (e.g. the keyboard proxy pointing to a display's
embedded keyboard chunk).
- all embedded chunks of appropriate type are listed as options when
selecting a value for a Chunk property.

If you have any config files that used the new 1.1 keyboard-input-from-display
window feature, they are now broken.  to fix, load into vjcontrol and
reselect the value in the keyboard proxy chunk.

Caveats:
 - the individual names of embedded chunks are editable in vjcontrol only
if the chunk is big enough that it gets edited in a separate window by
vjcontrol.  the next phase of vjcontrol renovations will add a flag to
make that fully customizable.
 - dependency checking in vjcontrol and juggler's dependency sort may not
deal with multiple levels of embedding correctly.  to be fixed tomorrow.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@4810 08b38cba-cd3b-11de-854e-f91c5b6e4272