This check-in adds support for block-io statistics collections. This adds a new directory vpr/IO/Stats. The directory contains a base stat collection strategy and a couple of other strategies as well as a template adapter that composes multiple strategies together into a single strategy.
To support this I also modified the current Socket_t and SocketStream_t classes to allow the passing of a template parameter that specifies the type of stat
collection strategy to use. Currently I have it defaulting to using a strategy that collects bandwidth information, but this could be changed in the future.

One interesting feature is that there is a NullStrategy that can be set.  When this strategy is set the compiler creates code that does not do anything more
then was done before.

NOTE: This is currently a work in progress and will definately need to be refectored more in the future.  (for example when we add latency collection support)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@5424 08b38cba-cd3b-11de-854e-f91c5b6e4272