Added boostTest test case.
The idea of this test is to test for all the behavior in boost that we require for our libraries to function correctly.

That way if boost changes in some way then we detect it here instead of seeing errors all over the place in our code.

This also provides a great place to experiment with boost features/abilities that you don't exactly know how to use yet. :)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@5644 08b38cba-cd3b-11de-854e-f91c5b6e4272