Tyring to clean up the destruction of Sockets.
- For now: Socket*_t bridge templates do NOT do anything for destruction.  Instead they rely upon the destructor of the member variable that
is of the actual implementation type.

Re-introduced some of the old SocketTest tests.  These tests failed because we didn't handle socket destruction correctly (and still don't
completely).  Adding them back in to test it now. (we should probably add in the other commented out tests soon too)

NOTE: There is an assertion still in this code.  I need a MT-aware debugger to fix it though, so I need to check it in broken.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@5649 08b38cba-cd3b-11de-854e-f91c5b6e4272