Optimizations to ComponentFactory.  Originally, it had a lot of lookup & caching in order to deal with all the classloaders it used.  But all that's been pointless since I introduced the unified vjClassLoader object, which handles class-loading for all vjcontrol extension .jars. The resulting code is much smaller & should be just as fast.
Modified classSupportsInterface query function so that it takes an
actual Class object as the interface argument instead of a string interface
name.  This saves a look up and protects us from stupid mistakes like
spelling errors.

Added a registerBaseJar() method that allows us to perform component
registration & configuration for stuff in the base vjcontrol.jar archive
without having to know the name & path for that .jar file.  This removes
an ugly filename dependency in VjControl.Init.

Whild playing with this, I tweaked a couple of methods' implementations
to be more friendly to VjComponents that happen to be aggregate javabeans.
Not that I expect bean aggregates to ever actually happen, but that's what
they said about Y2K too ;)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6242 08b38cba-cd3b-11de-854e-f91c5b6e4272