Created a VjComponent wrapper that encapsulates plexus' plxview network monitoring application.  This is not in any way a serious integration of the two components - there is no sharing of network capabilities, for example.  It is, however, a serious test of VjControl's componentization capabilities, and shows off some current shortcomings of the implementation:
- In order to use this component, you also have to load all the jar files
that make up plxview.  This leads to a tedious loading process or a
cumbersome command line like:
$JCCL_BASE_DIR/bin/vjcontrol \
-Djava.security.policy="$PLX_BASE_DIR/bin/policy.txt" \
-j $PLX_BASE_DIR/bin/openjgraph.jar \
-j $PLX_BASE_DIR/bin/PlexusGraphView.jar \
-j $PLX_BASE_DIR/bin/PlexusComm.jar \
-j $PLX_BASE_DIR/bin/PlxView.jar \
-j PlxViewComponent/PlxViewComponent.jar

yech.  It would be nice if the .jar containing the actual component could
tell VjControl to load all those others, but how should we deal with
locating the files?

- plxview needs a SecurityManager to exist in order to use certain RMI
features.  VjControl doesn't set one, so one of the plxview classes
creates an RMISecurityManager, which screws up dynamic loading on Irix
(it also shows that RMISecurityManager is implemented somewhat differently
on Irix v. linux, which is curious).  Can VjControl define a security
manager that doesn't interfere with any reasonable component?  What
restrictions can/should we place on code that is received via RMI or
a downloaded VjComponent?  Why the heck does java require a
SecurityManager for the RMI class loader but not the URL class loader,
anyway?


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6266 08b38cba-cd3b-11de-854e-f91c5b6e4272