Fixed documentation to jive with how ClassLoaders in VjControl are currently handled via the unified VjClassLoader object.
Marked add/removeComponentFactoryListener as deprecated; will probably
remove later.  In order for these events to have actually been useful,
we would need to send the event as soon as (but not before) a given class
was available for instantiation.  Unfortunately, VjControl rapidly outgrew
the point where that was practical.  When VjComponents can include
elements (such as other VjComponents or JavaBeans) in other .jar files,
there is no low-overhead way to ensure these messages are fired at a
reasonable time.  That's why this mechanism was deprecated in favor of
the ConfigChunk-based dynamic reconfiguration scheme implemented in
VjControl.Core.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6290 08b38cba-cd3b-11de-854e-f91c5b6e4272