Use the "AnalogData" class for storing & passing analog values.  as with the DigitalData checkin earlier today, these aren't currently taking advantage of the timestamping.
One possibly controversial change this required was moving normalization
from the devices' getAnalogData and pushing it up into sample and/or
udpateData.  We need to do this so that we can safely return the same
AnalogData from two consecutive getAnalogData calls without normalizing
the contained data twice, which is dangerous & silly.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6538 08b38cba-cd3b-11de-854e-f91c5b6e4272