More stuff figured out: Object activation within the POA now works.
The SubjectManager servant is registered with the POA using explicit
registration (which is the right technique since SubjectManagerImpl is
currently a singleton), and the reference created within the POA is
requested for binding in the naming service.  If all is working correctly,
we should be able to get references to the SubjectManager object from
client applications (e.g., a Java GUI).


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@6654 08b38cba-cd3b-11de-854e-f91c5b6e4272