Bug fixed:	Distributed deadlock between the observer and subject.
Make update() a one-way function.  This is needed so that omniORB does not
get stuck in a deadlcok when the subject invokes its observers' update()
methods as a result of the Java client calling the subject's notify()
method.  Tricky, huh?  I spent *all* *damn* day* trying to debug this only
to have it come down to a seven-character change.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@6701 08b38cba-cd3b-11de-854e-f91c5b6e4272