Bug fixed:	The commands given to be executed based on the version 		number check were only executed on the first pass (i.e., 		when the result of the check was not cached).
DPP_VERSION_CHECK_MSG() only executes the commands given to it when the
result of the check is not cached, and this leads to problems if callers
want to execute commands every time.  I am not sure if this is the best
way to solve this or if users should even be encouraged to execute commands
except when the result is not cached.  At this time, however, there is no
good way to inform the caller of the results of the version comparison
because the comparison is buried under several levels of macro expansion.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6739 08b38cba-cd3b-11de-854e-f91c5b6e4272