Bug fixed:	getAllSubjects() was not returning a valid sequence.
Apparently setting the size of a sequence in its constructor does not
work (or I did not do it correctly).  Instead, we'll est it explicitly so
that the sequence is sized correctly before elements are added.

I have also left in the debugging output that I used to track this one
down.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6760 08b38cba-cd3b-11de-854e-f91c5b6e4272