Bug fixed:	Detaching observers seems to work now.
Revert to using a std::vector<> of Observer_var objects.  I think I have
it working better now.  Detaching observers appears to work in the test
cases.  The observers are removed from the vector using remove_if() which
is supposed to be a safe way for performing such an operation.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@6771 08b38cba-cd3b-11de-854e-f91c5b6e4272