New feature:	VPR now has simulated sockets.
Sim Sockets Generation 2, take 2.

Instead of trying to force the use of sim sockets in Plexus instead of
real VPR sockets, the code has been moved into VPR.  This way, we can
build VPR in simulator mode and do whatever we want with types in
exactly the same way that we build the NSPR, POSIX, or Win32 subsystems.
This is a couple of orders of magnitude easier and cleaner.

To build VPR in simulator mode, use --enable-simulator when running
configure.  This will honor whatever thread subsystem is appropriate, but
it will replace all of the I/O and Interval code with simulator code.

VPR is now at version 0.6.0.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6833 08b38cba-cd3b-11de-854e-f91c5b6e4272