New feature:	Implemented simulated time.
The simulation controller can now manage the time of the simulation.  The
controller itself (vpr::sim::Controller) is a special thread-specific
singleton so that we may run multiple simulations in multiple threads.
The Socket Manager is no longer a singleton, and instead, an instance of
vpr::sim::SocketManager is accessed through the Controller singleton.

This change is what I was working towards all weekend.  Unfortunately, this
code (specifically, the Boost Graph Library part) does not seem to compile
on IRIX, so I am in a tough spot.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@6838 08b38cba-cd3b-11de-854e-f91c5b6e4272