Bug fixed:	Messages sent from acceptor to connetor (i.e., down the 		"reverse" path) never arrived.
Apparently I don't know how to use reverse iterators.  In an attempt to
improve performance, I wrote NetworkGraph::reversePath() that simply
reversed the given vertex list.  This was done to avoid calling
getShortestPath(u, v) and following it with getShortestPath(v, u).
However, my code was busted and caused an extra *incorrect* node to be
added at the front of the path.  A message sent on this path was then stuck
in limbo.  This was a nasty, nasty bug that had been preventing me from
making any progress since 1:00 am this morning.  :(


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6878 08b38cba-cd3b-11de-854e-f91c5b6e4272