Second phase of the zero copy simulator network.
- Added an extra special socket interface to the SIM sockets that takes shared_ptrs to data buffers.
- When running in simulator, use this interface directly to help performance.

Overall we get a 75% performance increase using this change, but there are still more tweaks to be made.

One bad thing of note.  This slowed down the message factory buildMessage function by 19%.  Mostly likely this is due to a new if statement that
I had to add to surround the entire buildMessage code.  Luckily this only affects the simulator message factory performance.  I am still trying
to think of ways to get rid of it though. I had thought about templates, but this did not work out so well because the function impl is in the
.cpp file and not in the .h.  I will look into it more soon though. (Maybe I will divide construction into two phases and template the first
phase....)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@6991 08b38cba-cd3b-11de-854e-f91c5b6e4272