Beginning of a fix for the new SocketManager.
The problem that was occuring (and still is) is because the design is based on the idea of each node having a unique IP address.  This is
not true though in the case of localhost.  In that single case a node that has a different IP address is being asked to act as 127.0.0.1 as
well.

One potential fix for this may be to treat 127.0.0.1 much like AnyAddr but just guarantee that all sockets with 127.0.0.1 will be mapped to
the *same* alternate address.  This could lead to some confusion when the address suddenly changes behind the back of the user, but
otherwise we have to litter special cases throughout the code.  This way we could just centralize it in one single place (that already
exists and does similar things).


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7049 08b38cba-cd3b-11de-854e-f91c5b6e4272