Bug fixed:	Sockets could not connect to each other.
Restored the ability to connect sockets.  This was broken in the Socket
Manager rewrite.  It happened mainly because of poorly chosen parameter
names and confusion about the use of pointers in the old code.  The
documentation explaining what was happening was correct, however.

The main change here is that the connector queue in SocketStreamImpSIM no
longer contains std::pair<> objects.  It was the *second* element of that
pair that needed to be removed, not the first.  This point alone was the
cause of a great deal of confusion.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7054 08b38cba-cd3b-11de-854e-f91c5b6e4272