Bug fixed:	Angled surface displays could cause assertion failures. SF Bug#:	469985
Use a tolerance-based comparison in calculateCornersInBaseFrame() to
deal with round-off error problems.  The observed error in testing was
less than 1e-6, so this is the value used as the tolerance.  This has
been tested on IRIX and Linux with displays rotated about the x axis and
about the y axis.


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@7130 08b38cba-cd3b-11de-854e-f91c5b6e4272