SF Bug#:	500148
Fixed compiling on Win32.  To deal with Visual C++ being unable to convert
unsigned 64-bit integers to double, we now cast to a signed 64-bit int and
then to double, but only on Win32.  setNowReal() has also been implemented
for Win32, but it seems as though it could have been implemented more
simply by using an NSPR call.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7252 08b38cba-cd3b-11de-854e-f91c5b6e4272