SF Bug#:	484422
Fixed the bug wherein the BeanTreeViewer's JTree would not refresh
properly if no Beans were loaded at initialization.  The cause of this
is not entirely clear to me yet, but this simple change seems to fix
it.  It may be worthwhile to revisit this in the future in case this
change is not the most optimal solution.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7253 08b38cba-cd3b-11de-854e-f91c5b6e4272