Bug fixed:	Any Panel Beans loaded at initialization and found after 		a Viewer Bean were not properly displayed in the viewer.
This big change fixes a nasty little bug that arose as a result of the
changes made when the BeanCollectionBuilder and ViewerRegistry were
removed.  (Actually, the bug was quite horrendous considering that it
ruined my whole day yesterday.)  Basically, the strict order in which
newly discovered Beans were handled was eliminated, and as a result, if
the list of .xml files returned by the file system was ordered such that
a Viewer Bean came before a Panel Bean, the Panel Bean would not show up
in the Viewer until that Viewer was refreshed forcibly.  The behavior
seen as a result of this bug was that sometimes Beans would show up, and
other times not.  Needless to say, that was very vexing.

This new implementation modifies the two existing Viewer Beans so that
they implement the new BeanRegistrationListener interface, and as a
result, they are informed automatically every time a new Bean is added to
the system (though they only care about Panel Beans).  In this way, there
is no longer a need for a manual refresh of the viewers after loading new
Beans, and that's definitely a good thing.  Since the need for the refresh
event is gone, I have removed the classes TreeModelRefeshEvent and
TreeModelRefreshListener.  They have been superceded by the
BeanRegistrationEvent-based code.  So, not only is the code less error
prone now, there is less of it to worry about.  :)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7261 08b38cba-cd3b-11de-854e-f91c5b6e4272