Bug fixed:	Improved handling of invalid enumeration strings. SF Bug#:	#506093
The config file parsers would get somewhat confused when a non-string
property had a value string that didn't match any of its enumeration
values (this could happen if the enumeration labels were changed,
for example).  This mostly affected old or hand-written config files,
as the writers generally use the value itself instead of the
enumeration label.

PropertyDesc.getEnumValue() now returns null correctly on a lookup
failure (matching the semantics of a java Map), and the parsers
check for this.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7354 08b38cba-cd3b-11de-854e-f91c5b6e4272