Act on the user-specified use of lazy Panel Bean instantiation.  This restores the original default behavior of the Tweek Java GUI while making it something users can turn on or off as they see fit.  All in all, I'd say that's a pretty positive change.
This change required some careful re-ordering of startup steps.  The order
is now:

    1. Register static Beans.
    2. Parse command-line arguments.
    3. Load global preferences.
    4. Find and load Beans in the Bean path.
    5. Start the GUI.

The Bean loading was moved from step 2 to step 4 so that the preferences
could be loaded before any Beans were discovered.  This also simplified
the code a little because there is now only one place that findAndLoadBeans()
is called.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7356 08b38cba-cd3b-11de-854e-f91c5b6e4272