Refactoring:	Removed dependency on custom code for sim camera projection drawing.
Modified viewports to have handles to parent displays,
and projections to have handles to parent viewports.
This enabled camera projection to be calculated
in CameraProjection class and to auto calculate the aspect ratio.
Now CameraProjection actually calculates a frutum so we do
not have to use a special cases for simulator
camera rendering where we use gluPerspective.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7460 08b38cba-cd3b-11de-854e-f91c5b6e4272