Bug fixed:	Backport of OSG app class and sample application didn't work at all.
This fixes the OSG app class and the sample application so that they work.

It is almost as if the developer that originally tried to do this was sick, had not slept in days, and had just taken a bunch of NyQuil.
(in fact I am sure this was the case)


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@7525 08b38cba-cd3b-11de-854e-f91c5b6e4272