New feature:	Context-specific file opening and closing.
This change implements context-specific (i.e., Bean-specific) file opening
and closing.  It is up to the Beans to do the actual file handling, but
the GUI is now capable of informing implementing Beans that the user wants
to open or close a file.  Part of this change includes the addition of a
"Close" item to the "File" menu.

Tweek is now at version 0.3.0.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7546 08b38cba-cd3b-11de-854e-f91c5b6e4272