Bug fixed:	A serial port could not be opened and closed repeatedly 		on Linux (and possibly other non-IRIX platforms).
When I originally wrote the serial port abstraction over a year ago, I
based it on the way that the existing VR Juggler device drivers were
written.  All of the drivers were written for IRIX and were presumably
based on some other driver code.  All of them zero out the serial port's
attributes immediately after opening.  I thought this was normal; it is
not.  On Linux (and possibly other non-IRIX platforms), this causes the
serial ports to get seriously messed up so that they cannot be re-opened
after being closed.

For the time being, I am leaving in the code that zeroes out the attributes
immediately after opening a port when building VPR on IRIX, but no other
platform will take those extra steps.  We will see for sure where it is
necessary, if it is even necessary at all.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7639 08b38cba-cd3b-11de-854e-f91c5b6e4272