New feature: Refactor: Bug fixed: SF Bug#:
Submitted by:
Reviewed by:
Approved by:
Obtained from:
Merge remote input branch into head branch.

In doing so, needed to make functions getDigitalData and and getPositionData
virtual to allow network data to be returned.  This was also a needed change
for device drivers in order for them to work with the recent input restructuring.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7651 08b38cba-cd3b-11de-854e-f91c5b6e4272