Converted the coding standard from the HTML book form (generated by install-web) to DocBook (using html2db and Morphon).  This version is scaled way down from the current version.  It removes the gigantic coding standard that we had downloaded from somwhere.  The appendix for the coding standard remains, but it will be filled in with our current, simple standard.  Besides that, I have removed conten that was duplicated between this and the Programmer's Guide.
This document is in pretty good shape for use with the VRJ 1.0 website.
We can probably publish it as-is and move on to updating it for all the
new stuff.

Using html2db drastically sped up the conversion process (the process took
a few hours hours instead of several days).  Because this document was
converted automatically, it has not received the thorough review that the
Programmer's Guide got.  People adding new content should bear that in
mind.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7710 08b38cba-cd3b-11de-854e-f91c5b6e4272